Move TweetNotSendException to Exceptions namespace
